Extract entire permissioned balances logic to separate file
